+++ src/Miner/Miner.h
@@ -47,7 +47,7 @@ private:
   Logging::LoggerRef m_logger;
 
   void runWorkers(BlockMiningParameters blockMiningParameters, size_t threadCount);
-  void workerFunc(const Block& blockTemplate, difficulty_type difficulty, uint32_t nonceStep);
+  void workerFunc(const Block& blockTemplate, difficulty_type difficulty, uint64_t nonceStep);
   bool setStateBlockFound();
 };
